+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -243,7 +243,7 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             drawerBackgroundSchemeColor: controller.drawerBackgroundSchemeColor,
             drawerWidth: controller.drawerWidth,
             drawerIndicatorWidth: controller.drawerIndicatorWidth,
-            drawerIndicatorBorderRadius: controller.drawerBorderRadius,
+            drawerIndicatorBorderRadius: controller.drawerIndicatorBorderRadius,
             drawerIndicatorSchemeColor: controller.drawerIndicatorSchemeColor,
             // BottomSheet settings.
             bottomSheetBackgroundColor: controller.bottomSheetSchemeColor,